requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.0
yfinance==0.2.28
schedule==1.2.0
python-dotenv==1.0.0